Add `size` property to `ArrayBuffer` native (#21)
* Add `size` property to `ArrayBuffer` class

The `size` property defines the total length of the memory block in bytes, making the buffer's range and usage clearer. Documentation and an example usage snippet were updated accordingly to reflect this change.

* Update default project version to 0.0.5-SNAPSHOT

Changed the fallback version from 0.0.3-SNAPSHOT to 0.0.5-SNAPSHOT in the build.gradle.kts file. This ensures consistency with the updated versioning strategy.